ARM: tegra: create common pre-dm i2c write
